Added support of Setting the Decimal Place
UnitDP default is 2, but you can pass in 4 when you get, create or
update BankTransactions, Invoices or Items.

You can set the DecimalPlace in the config.json -but this does not
automatically apply to your calls, you will still need to pass it using
config.getDecimalPlace() - I think a global 2 or 4 DP is not flexible
for those who may want more granular control.