Fix multi-object create
Xero's API support PUT for single and multiple objects on 10 endpoints.

- tried to use /endpoint POST for multiple and PUT for singluar (no id resource), but the POST resulted in an "update" and not create - therefore validation errors were not properly returned.

Added x-method extension to OpenAPI spec and set the value to PUT.  This is picked up in the mustache template if present and overrides the default httpMethod.

Confirmed that this solves the issue with summarizeError=false not functioning for multi object create.