Update tests for HistoryRecord component change
The OpenAPI spec we've moved HistoryRecord responses into components - this helps make the spec more DRY but means GetSandbox doesn't have an example response to return on all 13 "get" or "create" history endpoints.  The solution is to mock only the BatchPayment history get and create with an example response for use in the unit tests.  This validate ALL history calls since they all use the same part of the spec.

Remove Required attributes on objects that are used for create/update and as nested children.  This makes the objects are more flexibile for different use cases.  Objects changed.
- Items
- LinkedTransaction
- PurchaseOrder